+++ src/Service/LogoutService.php
@@ -8,7 +8,7 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Auth\Handler;
+namespace Aura\Auth\Service;
 
 use Aura\Auth\Adapter\AdapterInterface;
 use Aura\Auth\Session\SessionInterface;
@@ -22,7 +22,7 @@ use Aura\Auth\User;
  * @package Aura.Auth
  *
  */
-class LogoutHandler extends AbstractHandler
+class LogoutService extends AbstractService
 {
     /**
      *
