+++ src/Service/ResumeService.php
@@ -8,7 +8,7 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Auth\Handler;
+namespace Aura\Auth\Service;
 
 use Aura\Auth\User;
 use Aura\Auth\Adapter\AdapterInterface;
@@ -17,12 +17,12 @@ use Aura\Auth\Session\Timer;
 
 /**
  *
- * Resume Handler
+ * Resume Service
  *
  * @package Aura.Auth
  *
  */
-class ResumeHandler extends AbstractHandler
+class ResumeService extends AbstractService
 {
     protected $timer;
